Since left bitwise shift and overflow of signed integers can give undefined behavior, rework hitpoint and mana calculations for portability.  Avoids runtime error message from the undefined behavior sanitizer when Angband and the test cases are compiled with -fsanitize=undefined.  For similar portability concerns, use unsigned arithmetic when computing bit mask values.

+++ src/ui-prefs.c
@@ -364,7 +364,7 @@ void option_dump(ang_file *fff)
 			if (!window_flag_desc[j]) continue;
 
 			/* Only dump the flag if true */
-			if (window_flag[i] & (1L << j)) {
+			if (window_flag[i] & (((uint32_t) 1) << j)) {
 				file_putf(fff, "# Window '%s', Flag '%s'\n",
 						  angband_term_name[i], window_flag_desc[j]);
 				file_putf(fff, "window:%d:%d:1\n", i, j);
@@ -1049,9 +1049,9 @@ static enum parser_error parse_prefs_window(struct parser *p)
 	{
 		int value = parser_getuint(p, "value");
 		if (value)
-			d->window_flags[window] |= (1L << flag);
+			d->window_flags[window] |= (((uint32_t) 1) << flag);
 		else
-			d->window_flags[window] &= ~(1L << flag);
+			d->window_flags[window] &= ~(((uint32_t) 1) << flag);
 	}
 
 	d->loaded_window_flag[window] = true;
